Adds support for '@>' operator to Query language (#30).
The @> operator expects a property path expression on the lhs, and
a string literal that Postgresql can cast to a jsonb object.

On regular (registered) tables, the property path must reference a
column whose values can be cast to jsonb.

On Json tables, the property path will extract the property from the
json column, and cast it to jsonb.